requirements.txt
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
discord.py==2.3.2
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-logging==3.9.0
google-generativeai==0.3.2